Remove document v1 API endpoints
Fixes #7000
