+++ kuma/api/apps.py
@@ -12,5 +12,3 @@ class APIConfig(AppConfig):
 
     def ready(self):
         """Configure kuma.api after models are loaded."""
-        # Register signal handlers
-        from . import signal_handlers  # noqa
