+++ kuma/api/v1/tests/test_views.py
@@ -11,7 +11,6 @@ from stripe.error import APIError
 from waffle.models import Flag, Sample, Switch
 from waffle.testutils import override_flag
 
-from kuma.api.v1.views import document_api_data, get_content_based_redirect, get_s3_key
 from kuma.core.ga_tracking import (
     ACTION_SUBSCRIPTION_CANCELED,
     ACTION_SUBSCRIPTION_CREATED,
@@ -24,172 +23,6 @@ from kuma.search.tests import ElasticTestCase
 from kuma.users.models import UserSubscription
 from kuma.users.tests import user
 from kuma.wiki.models import BCSignal
-from kuma.wiki.templatetags.jinja_helpers import absolutify
-
-
-def test_get_s3_key(root_doc):
-    locale, slug = root_doc.locale, root_doc.slug
-    expected_key = "api/v1/doc/{}/{}.json".format(locale, slug)
-    assert get_s3_key(root_doc) == get_s3_key(locale=locale, slug=slug) == expected_key
-    assert (
-        get_s3_key(root_doc, prefix_with_forward_slash=True)
-        == get_s3_key(locale=locale, slug=slug, prefix_with_forward_slash=True)
-        == "/" + expected_key
-    )
-
-
-@pytest.mark.parametrize(
-    "case",
-    ("normal", "redirect", "redirect-to-self", "redirect-to-home", "redirect-to-wiki"),
-)
-def test_get_content_based_redirect(
-    root_doc, redirect_doc, redirect_to_self, redirect_to_home, redirect_to_macros, case
-):
-    if case == "normal":
-        doc = root_doc
-        expected = None
-    elif case == "redirect":
-        doc = redirect_doc
-        expected = (
-            get_s3_key(
-                root_doc, prefix_with_forward_slash=True, suffix_file_extension=False
-            ),
-            True,
-        )
-    elif case == "redirect-to-self":
-        doc = redirect_to_self
-        expected = None
-    elif case == "redirect-to-home":
-        doc = redirect_to_home
-        expected = ("/en-US/", False)
-    else:
-        doc = redirect_to_macros
-        expected = (absolutify("/en-US/dashboards/macros", for_wiki_site=True), False)
-    assert get_content_based_redirect(doc) == expected
-
-
-@pytest.mark.parametrize("http_method", ["put", "post", "delete", "options", "head"])
-def test_doc_api_disallowed_methods(client, http_method):
-    """HTTP methods other than GET are not allowed."""
-    url = reverse("api.v1.doc", args=["en-US", "Web/CSS"])
-    response = getattr(client, http_method)(url)
-    assert response.status_code == 405
-    assert_no_cache_header(response)
-
-
-def test_doc_api_404(client, root_doc):
-    """We get a 404 if we ask for a document that does not exist."""
-    url = reverse("api.v1.doc", args=["en-US", "NoSuchPage"])
-    response = client.get(url)
-    assert response.status_code == 404
-    assert_no_cache_header(response)
-
-
-def test_doc_api(client, trans_doc):
-    """On success we get document details in a JSON response."""
-    url = reverse("api.v1.doc", args=[trans_doc.locale, trans_doc.slug])
-    response = client.get(url)
-    assert response.status_code == 200
-    assert_no_cache_header(response)
-
-    data = response.json()
-    assert data["documentData"]
-    assert data["redirectURL"] is None
-    doc_data = data["documentData"]
-    assert doc_data["locale"] == trans_doc.locale
-    assert doc_data["slug"] == trans_doc.slug
-    assert doc_data["id"] == trans_doc.id
-    assert doc_data["title"] == trans_doc.title
-    assert doc_data["language"] == trans_doc.language
-    assert doc_data["hrefLang"] == "fr"
-    assert doc_data["absoluteURL"] == trans_doc.get_absolute_url()
-    assert doc_data["wikiURL"] == absolutify(
-        trans_doc.get_absolute_url(), for_wiki_site=True
-    )
-    assert doc_data["editURL"] == absolutify(
-        reverse("wiki.edit", args=(trans_doc.slug,), locale=trans_doc.locale),
-        for_wiki_site=True,
-    )
-    assert doc_data["translateURL"] is None
-    assert doc_data["bodyHTML"] == trans_doc.get_body_html()
-    assert doc_data["quickLinksHTML"] == trans_doc.get_quick_links_html()
-    assert doc_data["tocHTML"] == trans_doc.get_toc_html()
-    assert doc_data["translations"] == [
-        {
-            "locale": "en-US",
-            "language": "English (US)",
-            "hrefLang": "en",
-            "localizedLanguage": "Anglais am\u00e9ricain",
-            "title": "Root Document",
-            "url": "/en-US/docs/Root",
-        }
-    ]
-    assert doc_data["lastModified"] == "2017-04-14T12:20:00"
-
-
-def test_doc_api_for_redirect_to_doc(client, root_doc, redirect_doc):
-    """
-    Test the document API when we're requesting data for a document that
-    redirects to another document.
-    """
-    url = reverse("api.v1.doc", args=[redirect_doc.locale, redirect_doc.slug])
-    response = client.get(url, follow=True)
-    assert response.status_code == 200
-    assert_no_cache_header(response)
-
-    data = response.json()
-    assert data["documentData"]
-    assert data["redirectURL"] is None
-    doc_data = data["documentData"]
-    assert doc_data["locale"] == root_doc.locale
-    assert doc_data["slug"] == root_doc.slug
-    assert doc_data["id"] == root_doc.id
-    assert doc_data["title"] == root_doc.title
-    assert doc_data["language"] == root_doc.language
-    assert doc_data["hrefLang"] == "en"
-    assert doc_data["absoluteURL"] == root_doc.get_absolute_url()
-    assert doc_data["wikiURL"] == absolutify(
-        root_doc.get_absolute_url(), for_wiki_site=True
-    )
-    assert doc_data["translateURL"] == absolutify(
-        reverse("wiki.select_locale", args=(root_doc.slug,), locale=root_doc.locale,),
-        for_wiki_site=True,
-    )
-    assert doc_data["bodyHTML"] == root_doc.get_body_html()
-    assert doc_data["quickLinksHTML"] == root_doc.get_quick_links_html()
-    assert doc_data["tocHTML"] == root_doc.get_toc_html()
-    assert doc_data["translations"] == []
-    assert doc_data["lastModified"] == "2017-04-14T12:15:00"
-
-
-@pytest.mark.parametrize("case", ("redirect-to-home", "redirect-to-other"))
-def test_doc_api_for_redirect_to_non_doc(
-    client, redirect_to_home, redirect_to_macros, case
-):
-    """
-    Test the document API when we're requesting data for a document that
-    redirects to a non-document page (either the home page or another).
-    """
-    if case == "redirect-to-home":
-        doc = redirect_to_home
-        expected_redirect_url = "/en-US/"
-    else:
-        doc = redirect_to_macros
-        expected_redirect_url = absolutify(
-            "/en-US/dashboards/macros", for_wiki_site=True
-        )
-    url = reverse("api.v1.doc", args=[doc.locale, doc.slug])
-    response = client.get(url)
-    assert response.status_code == 200
-    assert_no_cache_header(response)
-
-    data = response.json()
-    assert data["documentData"] is None
-    assert data["redirectURL"] == expected_redirect_url
-
-    # Also ensure that we get exactly the same data by calling
-    # the document_api_data() function directly
-    assert data == document_api_data(redirect_url=expected_redirect_url)
 
 
 @pytest.mark.parametrize("http_method", ["put", "post", "delete", "options", "head"])
