+++ kuma/api/v1/urls.py
@@ -4,7 +4,6 @@ from . import views
 
 
 urlpatterns = [
-    re_path(r"^doc/(?P<locale>[^/]+)/(?P<slug>.*)$", views.doc, name="api.v1.doc"),
     re_path(r"^whoami/?$", views.whoami, name="api.v1.whoami"),
     re_path(r"^search/(?P<locale>[^/]+)/?$", views.search, name="api.v1.search"),
     re_path(r"^bc-signal/?$", views.bc_signal, name="api.v1.bc_signal"),
