+++ kuma/api/v1/views.py
@@ -9,12 +9,9 @@ from django.contrib.auth import get_user_model
 from django.http import (
     HttpResponse,
     HttpResponseBadRequest,
-    HttpResponsePermanentRedirect,
     JsonResponse,
 )
-from django.shortcuts import get_object_or_404
 from django.utils import translation
-from django.utils.translation import activate, gettext as _
 from django.views.decorators.cache import never_cache
 from django.views.decorators.csrf import csrf_exempt
 from django.views.decorators.http import require_GET, require_POST
@@ -54,183 +51,7 @@ from kuma.users.stripe_utils import (
     retrieve_and_synchronize_subscription_info,
 )
 from kuma.users.templatetags.jinja_helpers import get_avatar_url
-from kuma.wiki.models import Document
 from kuma.wiki.templatetags.jinja_helpers import absolutify
-
-
-@never_cache
-@require_GET
-def doc(request, locale, slug):
-    """
-    Return a JSON object that includes document content and metadata
-    for the document specified by the locale and path. Raises a 404
-    error if no such document exists. This is an API with URL
-    /api/v1/doc/<locale>/<path>
-    """
-    # TODO: This API endpoint probably needs to handle redirect documents
-    # and documents that fall back to the en-US locale. See
-    # the document() function in wiki/views/document.py for a model to follow.
-
-    # Since we don't have the locale at the start of the path, our
-    # locale middleware can't set the translation language correctly
-    # and we need to do it explicitly. (We need to know the language
-    # so that we can provide translated language names for the
-    # translations menu.)
-    activate(locale)
-    document = get_object_or_404(Document, locale=locale, slug=slug)
-
-    redirect = get_content_based_redirect(document)
-    if redirect:
-        redirect_url, is_redirect_to_document = redirect
-        if is_redirect_to_document:
-            return HttpResponsePermanentRedirect(redirect_url)
-        return JsonResponse(document_api_data(redirect_url=redirect_url))
-
-    return JsonResponse(document_api_data(document))
-
-
-def get_s3_key(
-    doc=None,
-    locale=None,
-    slug=None,
-    prefix_with_forward_slash=False,
-    suffix_file_extension=True,
-):
-    if doc:
-        locale, slug = doc.locale, doc.slug
-    key = reverse("api.v1.doc", args=(locale, slug))
-    if suffix_file_extension:
-        key += ".json"
-    if prefix_with_forward_slash:
-        # Redirects within an S3 bucket must be prefixed with "/".
-        return key
-    return key.lstrip("/")
-
-
-def get_cdn_key(locale, slug):
-    """Given a document's locale and slug, return the "key" for the CDN."""
-    return get_s3_key(
-        locale=locale,
-        slug=slug,
-        prefix_with_forward_slash=True,
-        suffix_file_extension=False,
-    )
-
-
-def get_content_based_redirect(document):
-    """
-    Returns None if the document is not a content-based redirect, otherwise a
-    tuple pair comprising the redirect URL as well as a boolean value. The
-    boolean value will be True if this is a redirect to another document,
-    otherwise False. If the document is a redirect to another document or a
-    redirect to the homepage, a relative URL will be returned, otherwise it
-    will be a full URL to the wiki site.
-    """
-    redirect_url = document.get_redirect_url()
-    if redirect_url and (redirect_url != document.get_absolute_url()):
-        redirect_document = document.get_redirect_document(id_only=False)
-        if redirect_document:
-            # This is a redirect to another document.
-            return (
-                get_s3_key(
-                    redirect_document,
-                    prefix_with_forward_slash=True,
-                    suffix_file_extension=False,
-                ),
-                True,
-            )
-        # This is a redirect to non-document page. For now, if it's the home
-        # page, return a relative path (so we stay on the read-only domain),
-        # otherwise return the full URL for the wiki site.
-        locale = document.locale
-        is_home_page = redirect_url in ("/", "/" + locale, "/{}/".format(locale))
-        if is_home_page:
-            # Let's return a relative URL to the home page for this locale.
-            return ("/{}/".format(locale), False)
-        # Otherwise, let's return a full URL to the Wiki site.
-        return (absolutify(redirect_url, for_wiki_site=True), False)
-    return None
-
-
-def document_api_data(doc=None, redirect_url=None):
-    """
-    Returns the JSON data for the document for the document API.
-    """
-    if redirect_url:
-        return {
-            "documentData": None,
-            "redirectURL": redirect_url,
-        }
-
-    # The original english slug for this document, for google analytics
-    if doc.locale == "en-US":
-        en_slug = doc.slug
-    elif doc.parent_id and doc.parent.locale == "en-US":
-        en_slug = doc.parent.slug
-    else:
-        en_slug = ""
-
-    other_translations = doc.get_other_translations(
-        fields=("locale", "slug", "title", "parent")
-    )
-    available_locales = {doc.locale} | set(t.locale for t in other_translations)
-
-    doc_absolute_url = doc.get_absolute_url()
-    revision = doc.current_or_latest_revision()
-    translation_status = None
-    if doc.parent_id and revision and revision.localization_in_progress:
-        translation_status = (
-            "outdated" if revision.translation_age >= 10 else "in-progress"
-        )
-    return {
-        "documentData": {
-            "locale": doc.locale,
-            "slug": doc.slug,
-            "enSlug": en_slug,
-            "id": doc.id,
-            "title": doc.title,
-            "summary": doc.get_summary_html(),
-            "language": doc.language,
-            "hrefLang": doc.get_hreflang(available_locales),
-            "absoluteURL": doc_absolute_url,
-            "wikiURL": absolutify(doc_absolute_url, for_wiki_site=True),
-            "editURL": absolutify(
-                reverse("wiki.edit", args=(doc.slug,), locale=doc.locale),
-                for_wiki_site=True,
-            ),
-            "translateURL": (
-                absolutify(
-                    reverse("wiki.select_locale", args=(doc.slug,), locale=doc.locale),
-                    for_wiki_site=True,
-                )
-                if doc.is_localizable
-                else None
-            ),
-            "translationStatus": translation_status,
-            "bodyHTML": doc.get_body_html(),
-            "quickLinksHTML": doc.get_quick_links_html(),
-            "tocHTML": doc.get_toc_html(),
-            "raw": doc.html,
-            "parents": [
-                {"url": d.get_absolute_url(), "title": d.title} for d in doc.parents
-            ],
-            "translations": [
-                {
-                    "language": t.language,
-                    "hrefLang": t.get_hreflang(available_locales),
-                    "localizedLanguage": _(settings.LOCALES[t.locale].english),
-                    "locale": t.locale,
-                    "url": t.get_absolute_url(),
-                    "title": t.title,
-                }
-                for t in other_translations
-            ],
-            "lastModified": (
-                doc.current_revision and doc.current_revision.created.isoformat()
-            ),
-        },
-        "redirectURL": None,
-    }
 
 
 @never_cache
