+++ kuma/settings/pytest.py
@@ -63,10 +63,6 @@ PIPELINE["COMPILERS"] = ("pipeline.compilers.sass.SASSCompiler",)
 WHITENOISE_AUTOREFRESH = True
 WHITENOISE_USE_FINDERS = True
 
-# This makes sure we our tests never actually use the real settings for
-# this.
-MDN_CLOUDFRONT_DISTRIBUTIONS = {}
-
 # Never rely on the .env
 GOOGLE_ANALYTICS_ACCOUNT = None
